[APDV-111] Added AddUnitConverterRequestBodyMapper
