+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/unit_converter/SaveUnitConverterImpl.java
@@ -2,7 +2,7 @@ package pl.edu.agh.apdvbackend.use_cases.unit_converter;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.edu.agh.apdvbackend.mappers.UnitConverterMapper;
+import pl.edu.agh.apdvbackend.mappers.unit.AddUnitConverterRequestBodyMapper;
 import pl.edu.agh.apdvbackend.mappers.unit.UnitConverterResponseBodyMapper;
 import pl.edu.agh.apdvbackend.models.body_models.unit_converter.AddUnitConverterRequestBody;
 import pl.edu.agh.apdvbackend.models.body_models.unit_converter.UnitConverterResponseBody;
@@ -13,14 +13,13 @@ import pl.edu.agh.apdvbackend.repositories.UnitConverterRepository;
 public class SaveUnitConverterImpl implements SaveUnitConverter {
 
     private final UnitConverterRepository unitConverterRepository;
-    private final UnitConverterMapper unitConverterMapper;
+    private final AddUnitConverterRequestBodyMapper requestBodyMapper;
     private final UnitConverterResponseBodyMapper responseBodyMapper;
 
     @Override
     public UnitConverterResponseBody execute(
             AddUnitConverterRequestBody addUnitConverterRequestBody) {
-        final var unitConverter = unitConverterMapper
-                .addRequestBodyToUnitConverter(addUnitConverterRequestBody);
+        final var unitConverter = requestBodyMapper.toUnitConverter(addUnitConverterRequestBody);
         final var savedUnitConverter = unitConverterRepository.save(unitConverter);
 
         return responseBodyMapper.toResponseBody(savedUnitConverter);
