+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/unit_converter/UpdateUnitConverterImpl.java
@@ -2,7 +2,7 @@ package pl.edu.agh.apdvbackend.use_cases.unit_converter;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.edu.agh.apdvbackend.mappers.UnitConverterMapper;
+import pl.edu.agh.apdvbackend.mappers.unit.AddUnitConverterRequestBodyMapper;
 import pl.edu.agh.apdvbackend.mappers.unit.UnitConverterResponseBodyMapper;
 import pl.edu.agh.apdvbackend.models.body_models.unit_converter.AddUnitConverterRequestBody;
 import pl.edu.agh.apdvbackend.models.body_models.unit_converter.UnitConverterResponseBody;
@@ -13,19 +13,18 @@ import pl.edu.agh.apdvbackend.repositories.UnitConverterRepository;
 public class UpdateUnitConverterImpl implements UpdateUnitConverter {
 
     private final UnitConverterRepository unitConverterRepository;
-    private final UnitConverterMapper unitConverterMapper;
+    private final AddUnitConverterRequestBodyMapper requestBodyMapper;
     private final UnitConverterResponseBodyMapper responseBodyMapper;
 
     @Override
     public UnitConverterResponseBody execute(
             Long unitConverterId,
-            AddUnitConverterRequestBody addUnitConverterRequestBody
+            AddUnitConverterRequestBody requestBody
     ) {
         final var unitConverter = unitConverterRepository
                 .findById(unitConverterId)
                 .orElseThrow();
-        final var updatedUnitConverter = unitConverterMapper
-                .updateAddRequestBodyToUnitConverter(addUnitConverterRequestBody, unitConverter);
+        final var updatedUnitConverter = requestBodyMapper.updateUnitConverter(requestBody, unitConverter);
 
         return responseBodyMapper.toResponseBody(unitConverterRepository.save(updatedUnitConverter));
     }
